\name{multibiplotGUI-package}
\alias{multibiplotGUI-package}
\docType{package}
\title{Multibiplot Analysis package}
\description{Provides a GUI with which users can construct and interact
        with Multibiplot Analysis}
\details{
\tabular{ll}{
Package: \tab multibiplotGUI\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2022-05-11\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Ana Belen Nieto Librero \email{ananieto@usal.es},
Nora Baccala \email{norabaccala@gmail.com},
Purificacion Vicente Galindo \email{purivg@usal.es},
Purificacion Galindo Villardon \email{pgalindo@usal.es}}

\keyword{ package }
